build: Move shared code into a package
An upcoming commit will add a command for performing rebuilds on
specific services, which will reuse a lot of the image building code
from the `up` command.
